Snap.Assets

Configuration

data Config

Constructors

Config

Fields

sourceDirectory :: String

The root path where the assets are kept. Paths you use in the builders are relative to this root. Most projects will have an 'assets/' directory where all the javascript, css, images are kept. The asset pipeline reads from that directory and serves those files to clients.

assetDefinitions :: [Asset]

The list of all assets that are known and served to clients. It is used by the snap handler and compiler to generate the files. If you have the snap handler disabled (eg. when running in production mode), it's safe to leave this list empty.

assetHost :: Snap String

The asset host used to serve the assets. Used by assetUrl and assetPath to generate the urls. Also see pathPrefix. You generally have two choices:

Extract the host from the request (default).

Use a single fixed host (eg. if you want to use a CDN).

The function should return the full authority, including the leading //. Scheme and port are optional.

 "//localhost"         -- Minimal definition.
 "//localhost:8000"    -- Non-standard port.
 "http://localhost"    -- Explicit scheme.

pathPrefix :: String

This prefix is added to paths generated by assetPath and assetUrl. Default is assets. If you have a dedicated asset host, you can leave this an empty string. The two most often used combinations of assetHost and pathPrefix are:

Using a dedicated asset host (or a CDN).

 config = { assetHost = return "assets.domain.tld", pathPrefix = "" }

   assetUrl   ->  "http://assets.domain.tld/file.js"
   assetPath  ->  "http://assets.domain.tld/file.js"

Serving assets from a subdirectory of the same host as your app.

 config = { assetHost = fromRequest, pathPrefix = "/assets" }

   assetUrl   ->  "http://domain.tld/assets/file.js"
   assetPath  ->  "/assets/file.js"

manifest :: Maybe Manifest

If set, asset urls and paths will be fingerprinted using the hashes stored in the manifest. This is something you'd want to enable when your application is running in production mode. You can store the manifest on disk (as YAML, JSON, plain text etc) and load when your application is starting, or compile it into your app.

type Manifest = Map String String

defaultConfig :: IO Config

The Asset type

data Asset

Describes an asset that is served to the clients. Each asset has a name (that is the string that is used in the incoming requests' URL) and a builder which specifies how to build the asset. The contents are basically a ByteString which is sent in the response back to the client.

Constructors

Asset

Fields

assetName :: String

assetBuilder :: Builder

Builders

type Builder

Arguments

= Config
Config so the builder knows where the root is

-> Maybe UTCTime
The contents of the If-Modified-Since header, so the builder can return NotModified if appropriate.

-> IO BuilderResult

Builds (compiles, generates, ..) the contents of an asset. Runs in the IO monad so you can do file IO etc. The builder should also have access to the Config so it knows where the root is.

data BuilderResult

The result of a build.

Constructors

NotModified
The contents have not been modified since the last request. This is especially important in development mode where you don't want to be needlessly rebuilding the assets.

Contents ByteString

concatBuilder :: [FilePath] -> Builder

Concatenate all files into a single blob.

Snap handlers

snapAssetHandler :: Config -> Snap ()

Template helpers

assetPath :: Config -> String -> String

assetUrl :: Config -> String -> Snap String

Generate the asset url for the asset with the given name. If we have a manifest we fingerprint the name with the hash from the manifest.

Tooling support

compileAssets :: Config -> String -> IO Manifest

For each asset, run the builder to generate the contents, then take the SHA1 of that and add to the manifest. Write the generated files into the output directory, using the correct (fingerprinted) name. The idea is then to upload the contents of the output directory to your CDN, so it can be served by fast/dedicated servers.